+++ classes/cohort_form.php
@@ -43,7 +43,8 @@ class cohort_form extends \moodleform {
         $mform = $this->_form;
         $cohorts = $this->_customdata['cohorts'];
 
-        $mform->addElement('html', \html_writer::tag('div', get_string('cohortsintro', 'local_profilecohort'),
+        $mform->addElement('html', \html_writer::tag('div', get_string('cohortsintro', 'local_profilecohort').'<br />'.
+                                                     get_string('invisiblecohortsnote', 'local_profilecohort'),
                                                      array('id' => 'intro', 'class' => 'box generalbox')));
 
         if (!$cohorts) {